Schemas

+++ src/models/skill.py
@@ -1,5 +1,5 @@
 from config.db_settings import Base
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Float, String, ForeignKey
 from sqlalchemy.orm import relationship
 
 
@@ -8,7 +8,7 @@ class Skill(Base):
 
     name = Column(String, primary_key=True, index=True)
     description = Column(String)
-    level = Column(Integer)
+    level = Column(Float)
     employee_username = Column(String, ForeignKey("employees.username"))
 
     employee = relationship("Employee", back_populates="skills")
